agents now handle directed SSH start messages
This includes adding a unique ID to the agent. Haven't worked out how other
components will discover this yet.

The alternative solution was to subscribe to `ssh.start` with a queue, so only
one sshark server picks up the message at a time.

A directed start is simpler for the receiving component, is less coupled to
the message bus, and allows us to implement placement logic similar to
`dea.advertise` in the future.